Fix WASM compiler stack imbalance, FASL loading diagnostics, and cold-boot-init
Three root-cause fixes that unblock the WASM32 root image build:

1. wasm2-immediate-set-xxx stack imbalance (compiler/WASM/wasm2.lisp):
   Unconditional local.get val-temp before if/else blocks left an
   unconsumed value in the store-ptr path, violating WASM stack
   discipline. Moved the push into the non-store-ptr branch only.
   Also fixed %setf-macptr to return the macptr object (not raw
   address), matching the ARM backend contract.

2. Cold-boot-init format call crash (level-0/nfasload.lisp):
   %wasm-note-startup-step called FORMAT when fboundp, but FORMAT
   needs CLOS stream infrastructure (SET-SLOT-ID-VALUE) which is
   only defined in level-1. Now v19+ boot modules validate FORMAT
   successfully, triggering this previously-latent bug. Removed the
   format call; C wrapper already tracks startup-step.

3. FASL loading infrastructure (kernel + subprims + JS):
   - Catch frame setup in wasm_fasload_path for proper error recovery
   - Funcall-nil detection with diagnostic logging
   - Fixed lisp_nil → nil_value for subprims module compatibility
   - Rich state dumps at funcall-error, ksignalerr, nthrow1value
   - JS-side diagnostics for non-zero faslRc
   - Fixed off-by-one in *fasl-api* diagnostic labels
   - Image build orchestration with cold-boot-init phases A/B/C

+++ lisp-kernel/wasm-kernel-request-smoke.c
@@ -21,9 +21,9 @@
 extern int lisp_open(char *path, int flags, mode_t mode);
 extern ssize_t lisp_read(int fd, void *buf, size_t count);
 extern ssize_t lisp_write(int fd, void *buf, size_t count);
-extern int64_t lisp_lseek(int fd, int64_t offset, int whence);
+extern int32_t lisp_lseek(int fd, int32_t offset, int whence);
 extern int lisp_close(int fd);
-extern int lisp_ftruncate(int fd, off_t length);
+extern int lisp_ftruncate(int fd, int32_t length);
 extern int lisp_stat(char *path, void *buf);
 
 __attribute__((used, visibility("default"), export_name("wasm_kernel_caps_abi_version")))
